[Refactor] Move TransactionFeeField into transaction folder.
