+++ src/features/Solana/Transaction/components/Main.tsx
@@ -9,8 +9,8 @@ import FieldDetail, { BadgeType } from "src/components/FieldDetail";
 import FormattedMessage from "src/components/FormattedMessage";
 import Header from "src/components/Header";
 import EstimatePointErrorField from "src/components/transaction/EstimatePointErrorField";
+import TransactionFeeField from "src/components/transaction/TransactionFeeField";
 import TransactionInfo from "src/components/transaction/TransactionInfo";
-import TransactionFeeField from "src/components/TransactionFeeField";
 import useEstimatePointInterval from "src/hooks/useEstimatePointInterval";
 import { useTransactionMachine } from "src/machines/transaction";
 import { logSendTx } from "src/services/Amplitude";
